extract server tracing tests
